Relative imports processed with futurized: RefactoringTool: Files that were modified: RefactoringTool: .\src\Selenium2Library\__init__.py RefactoringTool: .\src\Selenium2Library\webdrivermonkeypatches.py RefactoringTool: .\src\Selenium2Library\keywords\__init__.py RefactoringTool: .\src\Selenium2Library\keywords\_alert.py RefactoringTool: .\src\Selenium2Library\keywords\_browsermanagement.py RefactoringTool: .\src\Selenium2Library\keywords\_cookie.py RefactoringTool: .\src\Selenium2Library\keywords\_element.py RefactoringTool: .\src\Selenium2Library\keywords\_formelement.py RefactoringTool: .\src\Selenium2Library\keywords\_javascript.py RefactoringTool: .\src\Selenium2Library\keywords\_logging.py RefactoringTool: .\src\Selenium2Library\keywords\_runonfailure.py RefactoringTool: .\src\Selenium2Library\keywords\_screenshot.py RefactoringTool: .\src\Selenium2Library\keywords\_selectelement.py RefactoringTool: .\src\Selenium2Library\keywords\_tableelement.py RefactoringTool: .\src\Selenium2Library\keywords\_waiting.py RefactoringTool: .\src\Selenium2Library\locators\__init__.py RefactoringTool: .\src\Selenium2Library\locators\tableelementfinder.py RefactoringTool: .\src\Selenium2Library\utils\__init__.py RefactoringTool: .\src\Selenium2Library\utils\librarylistener.py RefactoringTool: .\src\Selenium2Library\utils\events\__init__.py RefactoringTool: .\src\Selenium2Library\utils\events\scope_event.py RefactoringTool: .\test\lib\mockito\__init__.py RefactoringTool: .\test\lib\mockito\inorder.py RefactoringTool: .\test\lib\mockito\invocation.py RefactoringTool: .\test\lib\mockito\mocking.py RefactoringTool: .\test\lib\mockito\mockito.py RefactoringTool: .\test\lib\mockito\spying.py
Imports in single lines and on top of files.
PEP8 E401 and E402
