+++ src/Selenium2Library/keywords/__init__.py
@@ -1,15 +1,16 @@
-from _logging import _LoggingKeywords
-from _runonfailure import _RunOnFailureKeywords
-from _browsermanagement import _BrowserManagementKeywords
-from _element import _ElementKeywords
-from _tableelement import _TableElementKeywords
-from _formelement import _FormElementKeywords
-from _selectelement import _SelectElementKeywords
-from _javascript import _JavaScriptKeywords
-from _cookie import _CookieKeywords
-from _screenshot import _ScreenshotKeywords
-from _waiting import _WaitingKeywords
-from _alert import _AlertKeywords
+from __future__ import absolute_import
+from ._logging import _LoggingKeywords
+from ._runonfailure import _RunOnFailureKeywords
+from ._browsermanagement import _BrowserManagementKeywords
+from ._element import _ElementKeywords
+from ._tableelement import _TableElementKeywords
+from ._formelement import _FormElementKeywords
+from ._selectelement import _SelectElementKeywords
+from ._javascript import _JavaScriptKeywords
+from ._cookie import _CookieKeywords
+from ._screenshot import _ScreenshotKeywords
+from ._waiting import _WaitingKeywords
+from ._alert import _AlertKeywords
 
 __all__ = [
     "_LoggingKeywords",
