+++ src/Selenium2Library/keywords/_alert.py
@@ -1,7 +1,8 @@
+from __future__ import absolute_import
 from selenium.common.exceptions import WebDriverException
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 
 class _AlertKeywords(KeywordGroup):
