+++ src/Selenium2Library/keywords/_cookie.py
@@ -1,4 +1,5 @@
-from keywordgroup import KeywordGroup
+from __future__ import absolute_import
+from .keywordgroup import KeywordGroup
 
 class _CookieKeywords(KeywordGroup):
