+++ src/Selenium2Library/keywords/_formelement.py
@@ -1,5 +1,6 @@
+from __future__ import absolute_import
 import os
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 from selenium.common.exceptions import WebDriverException
 
 class _FormElementKeywords(KeywordGroup):
