+++ src/Selenium2Library/keywords/_logging.py
@@ -1,7 +1,8 @@
+from __future__ import absolute_import
 import os
 import sys
 from robot.api import logger
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 from robot.libraries.BuiltIn import BuiltIn
 
 try:
