+++ src/Selenium2Library/keywords/_tableelement.py
@@ -1,7 +1,8 @@
+from __future__ import absolute_import
 import os
 import sys
 from Selenium2Library.locators import TableElementFinder
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 class _TableElementKeywords(KeywordGroup):
