+++ src/Selenium2Library/keywords/_waiting.py
@@ -1,6 +1,7 @@
+from __future__ import absolute_import
 import time
 import robot
-from keywordgroup import KeywordGroup
+from .keywordgroup import KeywordGroup
 
 class _WaitingKeywords(KeywordGroup):
