+++ src/Selenium2Library/locators/__init__.py
@@ -1,7 +1,8 @@
-from elementfinder import ElementFinder
-from tableelementfinder import TableElementFinder
-from windowmanager import WindowManager
-from customlocator import CustomLocator
+from __future__ import absolute_import
+from .elementfinder import ElementFinder
+from .tableelementfinder import TableElementFinder
+from .windowmanager import WindowManager
+from .customlocator import CustomLocator
 
 __all__ = [
     "ElementFinder",
