+++ src/Selenium2Library/locators/tableelementfinder.py
@@ -1,6 +1,7 @@
+from __future__ import absolute_import
 from selenium.common.exceptions import NoSuchElementException
 from Selenium2Library import utils
-from elementfinder import ElementFinder
+from .elementfinder import ElementFinder
 
 class TableElementFinder(object):
