+++ src/Selenium2Library/utils/__init__.py
@@ -1,8 +1,9 @@
+from __future__ import absolute_import
 import os
 from fnmatch import fnmatch
-from browsercache import BrowserCache
-from librarylistener import LibraryListener
-import events
+from .browsercache import BrowserCache
+from .librarylistener import LibraryListener
+from . import events
 
 __all__ = [
     "get_child_packages_in",
