+++ test/lib/mockito/invocation.py
@@ -1,7 +1,11 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-import matchers
+from __future__ import absolute_import
+#!/usr/bin/env python
+# coding: utf-8
+
+from . import matchers
 
 __copyright__ = "Copyright 2008-2010, Mockito Contributors"
 __license__ = "MIT"
