+++ test/lib/mockito/mockito.py
@@ -1,10 +1,14 @@
 #!/usr/bin/env python
 # coding: utf-8
 
-import verification
-from mocking import mock, TestDouble
-from mock_registry import mock_registry
-from verification import VerificationError
+from __future__ import absolute_import
+#!/usr/bin/env python
+# coding: utf-8
+
+from . import verification
+from .mocking import mock, TestDouble
+from .mock_registry import mock_registry
+from .verification import VerificationError
 
 __copyright__ = "Copyright 2008-2010, Mockito Contributors"
 __license__ = "MIT"
